Refactor error views and use svg logo text
