+++ gui/src/renderer/components/ErrorBoundary.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import styled from 'styled-components';
-import { colors, supportEmail } from '../../config.json';
+import { supportEmail } from '../../config.json';
 import { messages } from '../../shared/gettext';
 import log from '../../shared/logging';
 import PlatformWindowContainer from '../containers/PlatformWindowContainer';
-import { sourceSansPro } from './common-styles';
-import ImageView from './ImageView';
-import { Container, Layout } from './Layout';
+import ErrorView from './ErrorView';
 
 interface IProps {
   children?: React.ReactNode;
@@ -16,37 +14,8 @@ interface IState {
   hasError: boolean;
 }
 
-const StyledContainer = styled(Container)({
-  flex: 1,
-  flexDirection: 'column',
-  alignItems: 'center',
-  justifyContent: 'center',
-  backgroundColor: colors.blue,
-});
-
-const Title = styled.h1({
-  ...sourceSansPro,
-  fontSize: '26px',
-  lineHeight: '30px',
-  color: colors.white60,
-  marginBottom: '4px',
-});
-
-const Subtitle = styled.span({
-  fontFamily: 'Open Sans',
-  fontSize: '14px',
-  lineHeight: '20px',
-  color: colors.white40,
-  marginHorizontal: '22px',
-  textAlign: 'center',
-});
-
 const Email = styled.span({
   fontWeight: 900,
-});
-
-const Logo = styled(ImageView)({
-  marginBottom: '5px',
 });
 
 export default class ErrorBoundary extends React.Component<IProps, IState> {
@@ -75,13 +44,7 @@ export default class ErrorBoundary extends React.Component<IProps, IState> {
 
       return (
         <PlatformWindowContainer>
-          <Layout>
-            <StyledContainer>
-              <Logo height={106} width={106} source="logo-icon" />
-              <Title>MULLVAD VPN</Title>
-              <Subtitle role="alert">{reachBackMessage}</Subtitle>
-            </StyledContainer>
-          </Layout>
+          <ErrorView>{reachBackMessage}</ErrorView>
         </PlatformWindowContainer>
       );
     } else {
